requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==4.1.0
celery==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
cryptography==40.0.1
defusedxml==0.7.1
Django==4.1.7
django-celery-beat==2.5.0
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-modeltranslation==0.18.9
django-rosetta==0.9.8
django-timezone-field==5.0
djangorestframework==3.14.0
idna==3.4
kombu==5.3.1
oauthlib==3.2.2
Pillow==9.4.0
polib==1.2.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-crontab==2.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.1
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
